Split up utils.go into smaller files based on functionality grouping
